+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Mountain map convergence".
 - Tests for "Reverse FizzBuzz".
 - Tests for "Logic gates".
+- Tests for "Master of mayhem".
 
 ### Changed
 - Docker Node image from "14.16.1-alpine" to "node:16.14.2-alpine".